Fix tar archive layout
Opam package tar archive need a "toplevel" folder to be accepted by opam.
However, the previous way to implement this was assuming that the last folder of
`prefix` was "_opam", which might not be the case.

We now copy all needed files into a temporary "root" folder and use this
hierarchy to generate the tar archive.

Signed-off-by: Paul-Elliot <peada@free.fr>