fix version for `ocaml-system` sandbox pkg, keep original `ocaml`
`ocaml-system` needs to have the same version as `Sys.ocaml_version`. The
documentation says that this value will keep the `+` and `~` suffixes, but
experiments show that for the package `ocaml-variants.4.14.0~rc2+options`, the
value of `Sys.ocaml_version` is `4.14.0~rc2`. So we prefer to disable the
version check.
With the test disabled, we are no longer refrained from using the same package
for all suffixed version, by unsuffixing the version of `ocaml-system`. This
also allows us to not patch the `ocaml` package.

Signed-off-by: Paul-Elliot <peada@free.fr>