consider all tools binary only
As discussed in #128, distinguishing between tools that ship libraries and tools
that only consists of binaries has now more disadvantages than advantages:
Indeed, more tools will be added as dependencies of packages with the opam 2.2
`with-dev-tools`, and would be reinstalled by `opam`.

Although the new cache is compatible, it would be nice to have it versioned
before releasing this modification.

Now irrelevant tests are removed from the codebase.

Signed-off-by: Paul-Elliot <peada@free.fr>